refactor(ui): Extract a function to build the plugin template request body
