+++ ui/src/routes/admin/plugins/$pluginType/$pluginId/create-template/index.tsx
@@ -30,7 +30,7 @@ import { ChangeEvent } from 'react';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 
-import { PluginOption, PluginOptionTemplate } from '@/api';
+import { PluginOption } from '@/api';
 import { createPluginTemplateMutation } from '@/api/@tanstack/react-query.gen';
 import { OptionalInput } from '@/components/form/optional-input';
 import { ToastError } from '@/components/toast-error';
@@ -59,7 +59,10 @@ import { Switch } from '@/components/ui/switch';
 import { ApiError } from '@/lib/api-error';
 import { toast, toastError } from '@/lib/toast';
 import { getPluginTypeLabel } from '@/lib/types';
-import { buildPluginOptionsFormShape } from '../-helpers.ts';
+import {
+  buildPluginOptionsFormShape,
+  buildPluginTemplateRequestBody,
+} from '../-helpers.ts';
 import { Route as LayoutRoute } from '../../../route.tsx';
 
 const templateName = 'Template Name';
@@ -136,35 +139,10 @@ const CreateTemplate = () => {
 
   async function onSubmit(values: z.infer<typeof formSchema>) {
     const formValues = values as FormValues;
-    const requestBody: PluginOptionTemplate[] =
-      plugin?.options
-        ?.filter((option) => !formValues[`${option.name}_isNotSet`])
-        ?.map((option) => {
-          const value = formValues[option.name];
-          const isFinal = Boolean(formValues[`${option.name}_isFinal`]);
-
-          let stringValue: string | null;
-          if (value === null || value === undefined) {
-            stringValue = null;
-          } else if (Array.isArray(value)) {
-            stringValue = value.join(',');
-          } else if (
-            typeof value === 'bigint' ||
-            typeof value === 'number' ||
-            typeof value === 'boolean'
-          ) {
-            stringValue = value.toString();
-          } else {
-            stringValue = value as string;
-          }
-
-          return {
-            option: option.name,
-            type: option.type,
-            value: stringValue,
-            isFinal,
-          };
-        }) ?? [];
+    const requestBody = buildPluginTemplateRequestBody(
+      plugin?.options,
+      formValues
+    );
 
     await createTemplate({
       path: {
